+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.2] - 2024-01-27
+
+### Fixed
+
+- LSP (standalone): Use `bufnr` passed into `lsp.start` function when
+  determining detached file name.
+
 ## [4.0.1] - 2024-01-27
 
 ### Fixed